Opt rounds reveal input (#20)
* fix rounds

* remove ALICE input

* opt reveal and input

* opt rounds

* not reveal pub_M

* remove unused code

* fix memory leak

* remove reveal iv

* batch open

* refine

* flx set nonce

* split prove and check handshake

* remove warning

* safe setup protocol

---------

+++ backend/check_zero.h
@@ -19,12 +19,12 @@ inline void check_zero(const block* blk, size_t length, int party) {
 template <typename IO>
 inline void check_zero(const Integer& input, int party) {
     if (party == ALICE) {
-        for (int i = 0; i < input.size(); i++) {
+        for (size_t i = 0; i < input.size(); i++) {
             (((ZKProver<IO>*)(ProtocolExecution::prot_exec))->ostriple->auth_helper->hash)
               .put_block(&input[i].bit, 1);
         }
     } else {
-        for (int i = 0; i < input.size(); i++) {
+        for (size_t i = 0; i < input.size(); i++) {
             (((ZKVerifier<IO>*)(ProtocolExecution::prot_exec))->ostriple->auth_helper->hash)
               .put_block(&input[i].bit, 1);
         }
@@ -39,7 +39,7 @@ inline void check_zero(const Integer& input, const T* data, size_t len, int part
         error("inconsistent length!\n");
     bool* tmp = new bool[input.size()];
     if (party == ALICE) {
-        for (int i = 0; i < input.size(); i++)
+        for (size_t i = 0; i < input.size(); i++)
             tmp[i] = getLSB(input[i].bit);
 
         T* expected_data = new T[len];
