+++ backend/ole.h
@@ -19,7 +19,7 @@ class OLE {
         q = BN_new();
         BN_copy(this->q, q2);
         exp.resize(bit_length);
-        for (int i = 0; i < bit_length; ++i) {
+        for (size_t i = 0; i < bit_length; ++i) {
             exp[i] = BN_new();
             BN_set_bit(exp[i], i);
             BN_mod(exp[i], exp[i], q, ctx);
@@ -29,7 +29,7 @@ class OLE {
     ~OLE() {
         BN_CTX_free(ctx);
         BN_free(q);
-        for (int i = 0; i < bit_length; ++i)
+        for (size_t i = 0; i < bit_length; ++i)
             BN_free(exp[i]);
     }
 
@@ -40,9 +40,9 @@ class OLE {
         block* raw = new block[out.size() * bit_length];
         if (!cmpBlock(&ot->Delta, &zero_block, 1)) {
             ot->send_cot(raw, out.size() * bit_length);
-            for (int i = 0; i < out.size(); ++i) {
+            for (size_t i = 0; i < out.size(); ++i) {
                 BN_zero(out[i]);
-                for (int j = 0; j < bit_length; ++j) {
+                for (size_t j = 0; j < bit_length; ++j) {
                     H(pad1, raw[i * bit_length + j], q, ctx, ccrh);
                     H(pad2, raw[i * bit_length + j] ^ ot->Delta, q, ctx, ccrh);
                     BN_add(msg, pad1, pad2);
@@ -58,15 +58,15 @@ class OLE {
             }
         } else {
             bool* bits = new bool[out.size() * bit_length];
-            for (int i = 0; i < out.size(); ++i)
-                for (int j = 0; j < bit_length; ++j)
+            for (size_t i = 0; i < out.size(); ++i)
+                for (size_t j = 0; j < bit_length; ++j)
                     bits[i * bit_length + j] = (BN_is_bit_set(in[i], j) == 1);
 
             ot->recv_cot(raw, bits, out.size() * bit_length);
 
-            for (int i = 0; i < out.size(); ++i) {
+            for (size_t i = 0; i < out.size(); ++i) {
                 BN_zero(out[i]);
-                for (int j = 0; j < bit_length; ++j) {
+                for (size_t j = 0; j < bit_length; ++j) {
                     recv_bn(io, tmp);
 
                     H(msg, raw[i * bit_length + j], q, ctx, ccrh);
